% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIcodes.R
\name{CI.test}
\alias{CI.test}
\title{Perform conditional independence test for the variables X and Y conditional on Z, as described in Cai, Z., Li, R., & Zhang, Y. (2022).}
\usage{
CI.test(X, Y, Z, h)
}
\arguments{
\item{X}{A univariate variable}

\item{Y}{A univariate variable}

\item{Z}{A A univariate variable or multivariate vector}

\item{h}{kernel bandwidth}
}
\value{
The conditional indepenence measure.
}
\description{
Perform conditional independence test for the variables X and Y conditional on Z, as described in Cai, Z., Li, R., & Zhang, Y. (2022).
}
\examples{
n = 100
X1 = rnorm(n); X2 = rnorm(n); Z = rnorm(n)
X = X1+Z
Y = X2+Z
h = 1*1.06*sd(Z)*(4/(3*n))^{1/(1+4)}
CI.test(X, Y, Z, h)
}
