% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIcodes.R
\name{Independence.test}
\alias{Independence.test}
\title{Perform independence test for the variables X and Y using the method in Cai, Z., Li, R., & Zhang, Y. (2022).}
\usage{
Independence.test(X, Y)
}
\arguments{
\item{X}{A univariate variable}

\item{Y}{A univariate variable}
}
\value{
The indepenence measure.
}
\description{
Perform independence test for the variables X and Y using the method in Cai, Z., Li, R., & Zhang, Y. (2022).
}
\examples{
n = 100
X1 = rnorm(n); X2 = rnorm(n); Z = rnorm(n)
X = X1+Z
Y = X2+Z
Independence.test(X, Y)
}
